Use correct integer type

+++ src/include/stringfunctions.h
@@ -16,11 +16,11 @@
 namespace smash {
 
 /// Fill the string with characters to the left until the given width is reached.
-std::string fill_left(const std::string &s, int width, char fill = ' ');
+std::string fill_left(const std::string &s, size_t width, char fill = ' ');
 /// Fill the string with characters to the right until the given width is reached.
-std::string fill_right(const std::string &s, int width, char fill = ' ');
+std::string fill_right(const std::string &s, size_t width, char fill = ' ');
 /// Fill the string with characters at both sides until the given width is reached.
-std::string fill_both(const std::string &s, int width, char fill = ' ');
+std::string fill_both(const std::string &s, size_t width, char fill = ' ');
 
 /// takes a string and strips leading and trailing whitespaces.
 std::string trim(const std::string &s);
